refactor: extract update handlers from update adapters and scope `TransitEditorService` per update task
# Conflicts:
#	application/src/main/java/org/opentripplanner/updater/trip/gtfs/GtfsRealTimeTripUpdateAdapter.java
